+++ src/splunkParser.ts
@@ -70,7 +70,10 @@ export function returnCompletionItemfromJSON(context: vscode.ExtensionContext, f
             }
         }
         let hoverDocs;
-        if ( hasRelated ) { documentation = new vscode.MarkdownString(documentation + docString + "\n##### Related Commands:\n" + entry[related]); hoverDocs = documentation + docString + "\n##### Related Commands:\n" + entry[related]; }
+        if ( hasRelated ) { 
+            documentation = new vscode.MarkdownString(documentation + docString + "\n##### Related Commands:\n" + entry[related]);
+            hoverDocs = documentation + docString + "\n##### Related Commands:\n" + entry[related];
+        }
         if ( hasType ) { 
             detail = detail + '(' + entry[functionType].toLowerCase().replace(re, "_") + ') ';
             splunkCompletionItem.kind = vscode.CompletionItemKind.Method;
@@ -98,6 +101,34 @@ export function returnCompletionItemfromJSON(context: vscode.ExtensionContext, f
     };
 }
 
+function testNewline(docText: string, position: number) {
+    for (position; position > 0; position--) {
+        if (docText[position] === " " || docText[position] === "\t" || docText[position] === "[") {
+            continue;
+        }
+        else if (docText[position] === "\n" || docText[position] === "\r") {
+            return true;
+        }
+        else {
+            return false;
+        }
+    }
+}
+
+function getMainCommand(docText: string, position: number) {
+    //assuming we trigger this on every pipe or open bracket, we can assume
+    // the next word is a command
+    let word: string = '';
+    for (position; position > 0; position++) {
+        if ((docText[position] === " " || docText[position] === "\t" || docText[position] === "[" || docText[position] === "|") && word === "") {
+            continue;
+        }
+        else if (docText[position]) {
+            
+        }
+    }
+}
+
 export function returnCommandRegister(context: vscode.ExtensionContext) {
     const splunkCommandHandler = vscode.commands.registerCommand('splunk_search.Prettify', function() {
         let editor = vscode.window.activeTextEditor;
@@ -105,18 +136,22 @@ export function returnCommandRegister(context: vscode.ExtensionContext) {
             console.log("editor is active");
             let document = editor.document;
             
+            //docText is actual string we're going to return, docTextCopy is 
+            // a copy to iterate through that doesn't change
             let docText: string = document.getText();
             let docTextCopy: string = docText;
 
             // define new vars:
-            // keep track of position in string
+            // keep track of position in docText
             let i: number = 0; 
 
-            // keep track of whether or not we're in a bracket
+            // keep track of whether or not we're going to skip a pipe
             let inBaseBracket: boolean = false;
 
             // keep track of what level we're at in brackets
             let bracketLevel = 0;
+
+            // keep track of what global splunk command we're in 
             for (const character of docTextCopy) {
                 if (i === 0) { i++; continue; }
                 if (inBaseBracket) {
@@ -137,6 +172,11 @@ export function returnCommandRegister(context: vscode.ExtensionContext) {
                 }
                 else if (character === "[") {
                     console.log("open bracket found");
+                    if (testNewline(docText, i-1)) {
+                        console.log("found newline with function");
+                        i++;
+                        continue;
+                    }
                     let lengthToIncrease: number = 2;
                     let insertString: string = '\n';
                     
@@ -149,9 +189,9 @@ export function returnCommandRegister(context: vscode.ExtensionContext) {
                 }
                 else if (character === "]") {
                     console.log("close bracket found");
-                    let lengthToIncrease: number = 2;
-                    let insertString: string = '\n';
+                    let lengthToIncrease: number = 1;
                     
+                    let insertString: string = '';
                     // make sure bracketLevel is decreased before adding in \t
                     bracketLevel--;
                     for (var x = 0; x < bracketLevel; x++) { insertString = insertString + '  '; lengthToIncrease += 2; }
@@ -160,11 +200,17 @@ export function returnCommandRegister(context: vscode.ExtensionContext) {
                 }
                 else if (character === '|') {
                     console.log("Pipe found");
-                    console.log(docText[i-1]);
-                    if (docText[i] === "\n" || docText[i - 1] === "\n" || docText[i - 2] === "\n") {
+                    
+                    if (testNewline(docText, i-1)) {
+                        console.log("found newline with function");
                         i++;
                         continue;
                     }
+                    /* if (docText[i] === "\n" || docText[i - 1] === "\n" || docText[i - 2] === "\n") {
+                        i++;
+                        console.log("found newline; leaving it alone");
+                        continue;
+                    } */
                     let lengthToIncrease: number = 2;
                     let insertString: string = '\n';
                     
@@ -177,13 +223,11 @@ export function returnCommandRegister(context: vscode.ExtensionContext) {
                     i++;
                 }
             }
-            console.log(docText);
+            
             editor.edit(editBuilder => {
                 let numLines: number = document.lineCount;
                 console.log(numLines);
                 let range: vscode.Range = new vscode.Range(0, 0, numLines, document.lineAt(numLines - 1).text.length);
-                console.log(range);
-                console.log(range.start, range.end);
                 editBuilder.replace(range, docText);
                 console.log("replaced");
             });