+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to the "vscode-splunk-search-linter" extension will be docum
 
 ## [Unreleased]
 
+## [0.4.0] - 2020-02-06
+
+### Added
+
+- `Format SPL` VSCode command to format searches
+- Hover commands and keywords
+
 ## [0.3.0] - 2020-01-22
 
 ### Changed